Creating migration for Dictionary -> CodeableData for the redisplay dict
If the OS_IAM_REDISPLAY_CACHE is successfully read as Codeable data we don't need to migrate. If that fails try to read it as a dictionary and then save it as Codeable data. If that also fails then clear the cache because we cannot determine what data type was stored for the key.